Teach image comparison to handle width differences.
Although the standard SNES video mode is 256 pixels wide, the SNES is
capable of 512px output, and the underlying video-hardware actually
outputs two identical pixels for each input pixel when not in hi-res
mode - so it's perfectly legitimate for a SNES emulator to output 512px
lines all the time, not taking advantage of the common optimization of
outputting 256 pixels in low-res modes.

Therefore, if we are comparing an image from an always-hi-res core
versus a hi-res-when-needed core, it's not necessarily a problem that
they have different dimensions. The image_difference() method has been
adjusted to take this corner-case into account.