Support bsnes' XML shader format.
Adds support for GLSL shaders in bsnes' XML shader files to the
glut-demo script. Just pass the path to a shader file as the second
argument (the first is the path to the ROM image to load) and it will be
compiled and applied to the scene. If there's a compile error, the error
message will be displayed and the script will exit without doing any
emulation.

I'll probably need to refactor a lot of the code from bin/glut-demo into
gl_output.py, but that can come later.